% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Aggregate_epiRegions}
\alias{Aggregate_epiRegions}
\title{Aggregate Bismark Coverage Files into Defined Genomic Regions}
\usage{
Aggregate_epiRegions(cov_dir, bed_file, n_cores = 1)
}
\arguments{
\item{cov_dir}{Character. Path to the directory containing `.cov.gz` files.}

\item{bed_file}{Character. Path to the BED file. Must contain at least 3 columns: chr, start, end.}

\item{n_cores}{Integer. Number of cores to use for parallel processing. Default is 1.}
}
\value{
A \code{data.table} containing the aggregated results. The first column is \code{region_id},
followed by \code{.meth}, \code{.nonmeth}, and \code{.level} columns for each sample.
}
\description{
This function takes a directory of Bismark `.cov.gz` files and a BED file containing
genomic regions. It aggregates the methylation data across these regions in parallel.
}
