% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc_variance.R
\name{Chr_region_process}
\alias{Chr_region_process}
\title{Distinguish chrXX:xxx-xxx as chrXX, xxx, xxx}
\usage{
Chr_region_process(file_tmp, method)
}
\arguments{
\item{file_tmp}{chromosome physical segment data, chrXX:xxx-xxx format or chrXX, xxx, xxx format}

\item{method}{chromosome fragment processing method}
}
\value{
Processed physical fragment data, chrXX, xxx, xxx format or chrXX:xxx-xxx format
}
\description{
Distinguish chrXX:xxx-xxx as chrXX, xxx, xxx
}
