% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Cov_QC}
\alias{Cov_QC}
\title{Filtering of site information}
\usage{
Cov_QC(files, save_path)
}
\arguments{
\item{files}{Site information file path}

\item{save_path}{Save folder path}
}
\value{
nothing
}
\description{
Filtering of site information
}
