% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Cov_to_data}
\alias{Cov_to_data}
\title{Cov files to text files}
\usage{
Cov_to_data(cov_file, cov_file_data, chr_tmp, region, suffixname, datatmp)
}
\arguments{
\item{cov_file}{site file name}

\item{cov_file_data}{site data}

\item{chr_tmp}{chromosome data}

\item{region}{chromosomal physical fragment}

\item{suffixname}{file extension}

\item{datatmp}{select the meth or meth-level method}
}
\value{
DNA methylation or chromatin accessibility matrix
}
\description{
Cov files to text files
}
