% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Coverage_to_data}
\alias{Coverage_to_data}
\title{Coverage files to text files}
\usage{
Coverage_to_data(
  cov_file,
  cov_file_data,
  region_data,
  chr_data,
  suffixname_data = ".cov.gz",
  type
)
}
\arguments{
\item{cov_file}{site file name}

\item{cov_file_data}{site data}

\item{chr_data}{chromosome data}

\item{region}{chromosomal physical fragment}

\item{suffixname}{file extension}

\item{datatmp}{select the meth or meth-level method}
}
\value{
DNA methylation or chromatin accessibility matrix
}
\description{
Coverage files to text files
}
