% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{DEG_diff_plot}
\alias{DEG_diff_plot}
\title{Difference level chart}
\usage{
DEG_diff_plot(CpG_DEG_file_data, GpC_DEG_file_data)
}
\arguments{
\item{CpG_DEG_file_data}{Methylation differences}

\item{GpC_DEG_file_data}{Open difference information}
}
\value{
CpG and GpC differences data
}
\description{
Difference level chart
}
