% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{DR_process}
\alias{DR_process}
\title{Dimensionality reduction analysis}
\usage{
DR_process(
  file_data,
  method = "RNA_UMAP",
  n_neighbors = 100,
  min_dist = 0.5,
  center = T,
  scale = T,
  cor_method = "pearson",
  rank. = 30,
  metric = "cosine"
)
}
\arguments{
\item{file_data}{Transposed expression matrix}

\item{method}{Behavioral cells, classified as genes}

\item{n_neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation.}

\item{min_dist}{The effective minimum distance between embedded points.}

\item{center}{a logical value indicating whether the variables should be shifted to be zero centered.}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{rank.}{optionally, a number specifying the maximal rank, i.e., maximal number of principal components to be used.}

\item{metric}{Type of distance metric to use to find nearest neighbors}
}
\value{
Dimensionality-Reduced Dataset
}
\description{
Dimensionality reduction analysis
}
