% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{DReduction}
\alias{DReduction}
\title{Dimensionality reduction analysis}
\usage{
DReduction(file_data, method = "UMAP")
}
\arguments{
\item{file_data}{Transposed expression matrix}

\item{method}{Behavioral cells, classified as genes}
}
\value{
Dimensionality-Reduced Dataset
}
\description{
Dimensionality reduction analysis
}
