% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{Group_model_predict}
\alias{Group_model_predict}
\title{Grouping model prediction}
\usage{
Group_model_predict(file_tmp, file_group, test_size = 0.3, model = "lgbm")
}
\arguments{
\item{file_tmp}{Enter the RNA expression matrix}

\item{file_group}{Grouping information for the sample}

\item{test_size}{Training-to-testing ratio}

\item{model}{Model selection}
}
\value{
Post-training data
}
\description{
Grouping model prediction
}
