% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_padding_process.R
\name{KNN_padding}
\alias{KNN_padding}
\title{The NA value is populated by the KNN}
\usage{
KNN_padding(file_tmp, sample_cutoff = 0.8)
}
\arguments{
\item{file_tmp}{Methlevel data of DNA methylation or chromatin accessibility}

\item{sample_cutoff}{column cutoff}
}
\value{
After filling the matrix
}
\description{
The NA value is populated by the KNN
}
