% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{LASSO_feature}
\alias{LASSO_feature}
\title{LASSO gene}
\usage{
LASSO_feature(file_tmp, file_group_tmp, test_size = 0.3)
}
\arguments{
\item{file_tmp}{Enter the RNA expression matrix}

\item{file_group_tmp}{Grouping information for the sample}

\item{test_size}{Training-to-testing ratio}
}
\value{
Returns the expression matrix after selection
}
\description{
LASSO gene
}
