% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Methlevel_diff}
\alias{Methlevel_diff}
\title{Differential level data processing}
\usage{
Methlevel_diff(
  methlevel_data,
  group_data,
  group_suff,
  methlevel_group_suff,
  suff
)
}
\arguments{
\item{methlevel_data}{Chromosome-level locus data}

\item{group_data}{group data}

\item{group_suff}{grouping field}

\item{methlevel_group_suff}{methlevel grouping field}

\item{suff}{group information}
}
\value{
methlevel results
}
\description{
Differential level data processing
}
