% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc_variance.R
\name{Methlevel_group_variance_analysis}
\alias{Methlevel_group_variance_analysis}
\title{Methlevel group variance analysis}
\usage{
Methlevel_group_variance_analysis(
  methlevel_data,
  group_data,
  group_suff = "group",
  methlevel_group_suff = "methlevel",
  suff
)
}
\arguments{
\item{methlevel_data}{DNA methylation or chromatin accessibility methlevel data}

\item{group_data}{DNA methylation or chromatin accessibility group data}

\item{group_suff}{grouping field}

\item{methlevel_group_suff}{methlevel grouping field}

\item{suff}{grouping suffix name}
}
\value{
methlevel Difference analysis results
}
\description{
Methlevel group variance analysis
}
