% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_padding_process.R
\name{NA_padding_num_mean}
\alias{NA_padding_num_mean}
\title{The NA value is populated by the specified value and by the average value}
\usage{
NA_padding_num_mean(file_tmp, num = 0.5, flag = "row")
}
\arguments{
\item{file_tmp}{level data of DNA methylation or chromatin accessibility}

\item{num}{Enter the number to fill}

\item{flag}{Choose to count by column or row}
}
\value{
After filling the matrix
}
\description{
The NA value is populated by the specified value and by the average value
}
