% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{Pseudotime_analysis}
\alias{Pseudotime_analysis}
\title{UMAP dimensionality reduction data was used for quasi-temporal analysis}
\usage{
Pseudotime_analysis(rds_tmp)
}
\arguments{
\item{rds_tmp}{RDS objection}
}
\value{
monocle3 object and Pseudotime data
}
\description{
UMAP dimensionality reduction data was used for quasi-temporal analysis
}
