% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{QC_data}
\alias{QC_data}
\title{Methlevel quality control}
\usage{
QC_data(file_data, sample_cutoff = 0.8)
}
\arguments{
\item{file_data}{Data requiring column quality control}

\item{sample_cutoff}{Percentage of NA Values}
}
\value{
Data after quality control
}
\description{
Methlevel quality control
}
