% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{QC_methlevel_data}
\alias{QC_methlevel_data}
\title{Methlevel quality control}
\usage{
QC_methlevel_data(file_data, sample_cutoff = 0.8)
}
\arguments{
\item{sample_cutoff}{}
}
\description{
Methlevel quality control
}
