% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{Read_file_colname}
\alias{Read_file_colname}
\title{Read meth, UNmeth, and meth_level data, the corresponding column name is \\.name$}
\usage{
Read_file_colname(file_tmp, string_data)
}
\arguments{
\item{file_tmp}{DNA methylation or chromatin accessibility matrix}

\item{string_data}{sample type information, such as meth, UNmeth, methlevel}
}
\value{
sample data after selecting columns
}
\description{
Read meth, UNmeth, and meth_level data, the corresponding column name is \\.name$
}
