% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{RNA_data_standard}
\alias{RNA_data_standard}
\title{RNA standard data}
\usage{
RNA_data_standard(file_tmp, file_tmp_length, method = "tpm")
}
\arguments{
\item{file_tmp}{Enter the RNA expression matrix}

\item{file_tmp_length}{The length of each gene entered}

\item{method}{Methods for data standardization, TPM, FPKM, CPM}
}
\value{
Expression matrix after normalization
}
\description{
RNA standard data
}
