% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{SC_process}
\alias{SC_process}
\title{Calculate the silhouette coefficient}
\usage{
SC_process(file_data, file_type = "UMAP")
}
\arguments{
\item{file_data}{Dimensionality reduction data}

\item{file_type}{Dimensionality-Reduced data types}
}
\value{
Mean Silhouette coefficient value
}
\description{
Calculate the silhouette coefficient
}
