% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{CreatRNAObject}
\alias{CreatRNAObject}
\title{Create an RNA storage object}
\usage{
CreatRNAObject(file_data, min_cells = 50, min_features = 0)
}
\arguments{
\item{file_data}{RNA Expression Matrix}

\item{min_cells}{Include features detected in at least this many cells.}

\item{min_features}{Include cells where at least this many features are detected}
}
\value{
RNA object
}
\description{
Create an RNA storage object
}
