% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc.R
\name{DR_process}
\alias{DR_process}
\title{Dimensionality reduction processing}
\usage{
DR_process(file_data, method = "MDS")
}
\arguments{
\item{file_data}{Enter level data}

\item{method}{Dimension reduction methods}
}
\value{
Dimension reduction results
}
\description{
Dimensionality reduction processing
}
