% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_padding_process.R
\name{KNN_padding}
\alias{KNN_padding}
\title{The NA value is populated by the KNN}
\usage{
KNN_padding(file_tmp, k_data = 10)
}
\arguments{
\item{file_tmp}{level data of DNA methylation or chromatin accessibility}

\item{k_data}{Number of neighbors to be used in the imputation (default=10)}
}
\value{
After filling the matrix
}
\description{
The NA value is populated by the KNN
}
