% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc_variance.R
\name{Level_group_variance_analysis}
\alias{Level_group_variance_analysis}
\title{level group variance analysis}
\usage{
Level_group_variance_analysis(
  methlevel_data,
  group_data,
  group_suff = "group",
  methlevel_group_suff = "level",
  suff1,
  suff2 = NULL
)
}
\arguments{
\item{methlevel_data}{DNA methylation or chromatin accessibility methlevel data}

\item{group_data}{DNA methylation or chromatin accessibility group data}

\item{group_suff}{grouping field}

\item{methlevel_group_suff}{level grouping field}

\item{suff1}{grouping suffix name1}

\item{suff2}{grouping suffix name2}
}
\value{
methlevel Difference analysis results
}
\description{
level group variance analysis
}
