% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_acc_variance.R
\name{Meth_group_variance_analysis}
\alias{Meth_group_variance_analysis}
\title{Meth group variance analysis}
\usage{
Meth_group_variance_analysis(
  meth_data,
  UNmeth_data,
  group_data,
  group_suff = "group",
  meth_title_suff = "site",
  UNmeth_title_suff = "UNsite",
  suff
)
}
\arguments{
\item{meth_data}{DNA methylation or chromatin accessibility meth data}

\item{UNmeth_data}{DNA methylation or chromatin accessibility UNmeth data}

\item{group_data}{DNA methylation or chromatin accessibility group data}

\item{group_suff}{grouping field}

\item{meth_title_suff}{site grouping field}

\item{UNmeth_title_suff}{UNsite grouping field}

\item{suff}{group suffix name}
}
\value{
meth differential data
}
\description{
Meth group variance analysis
}
