% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_padding_process.R
\name{NA_count_order}
\alias{NA_count_order}
\title{count the number of NAs in the ranks}
\usage{
NA_count_order(file_tmp, flag = "row")
}
\arguments{
\item{file_tmp}{Methlevel data of DNA methylation or chromatin accessibility}

\item{flag}{Choose to count by column or row}
}
\value{
statistical information
}
\description{
count the number of NAs in the ranks
}
