% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_padding_process.R
\name{NA_padding_mean}
\alias{NA_padding_mean}
\title{The NA value is populated by the average}
\usage{
NA_padding_mean(file_tmp, flag = "row")
}
\arguments{
\item{file_tmp}{level data of DNA methylation or chromatin accessibility}

\item{flag}{Select a column or row}
}
\value{
After filling the matrix
}
\description{
The NA value is populated by the average
}
